requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
Django==5.1
django-cors-headers==4.4.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.7
mysqlclient==2.2.4
ndg-httpsclient==0.5.1
pusher==3.3.2
pyasn1==0.6.0
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==24.2.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
tzdata==2024.1
urllib3==2.2.2
